+++ src/lib/geojson.js
@@ -1,14 +1,15 @@
 const _ = require('highland')
 
-module.exports = {
-  createReadStream: function (input, options) {
-   const start = '{"type":"FeatureCollection","features":['
-   const end = ']}'
-   const features = options && options.json ? _(input).compact().map(JSON.stringify) : _(input).compact()
-   const readStream = _([start])
+function createStream (options) {
+  const start = '{"type":"FeatureCollection","features":['
+  const end = ']}'
+  const readStream = _.pipeline(s => {
+   const features = options && options.json ? _(s).compact().map(JSON.stringify) : _(s).compact()
+   return  _([start])
    .concat(features.intersperse(','))
    .append(end)
-
-   return readStream
- }
+  })
+  return readStream
 }
+
+module.exports = {createStream}