+++ src/lib/ogr.js
@@ -1,98 +1,36 @@
 const spawn = require('child_process').spawn
 const _ = require('highland')
 const sanitize = require('sanitize-filename')
+const fs = require('fs')
+const Shapefile = require('./shapefile')
+const Cmd = require('./ogr-cmd')
 
-const ogrFormats = {
-  kml: 'KML',
-  zip: '"ESRI Shapefile"',
-  csv: 'CSV',
-  json: 'GeoJSON',
-  geojson: 'GeoJSON',
-  gpkg: 'GPKG'
-}
+function createStream (format, options) {
+  options.input = `${options.path}/layer.vrt`
+  const vrt = fs.createWriteStream(options.input)
+  options.name = options.name ? sanitize(options.name) : 'output'
 
-module.exports = {
-  createReadStream: function (input, options) {
-    const readStream = _()
-    options.name = sanitize(options.name)
-    options.input = input
-    const cmd = this.createCmd(options)
+  const cmd = Cmd.create(format, options)
 
-    if (options.format === 'zip') return require('./shapefile').createReadStream(options)
-
-    const ogr = spawn('ogr2ogr', cmd)
-    ogr.stdout.on('data', data => readStream.write(data))
-    ogr.on('close', c => {
-      if (c > 0) readStream.emit('error', new Error('OGR Failed'))
-      else readStream.end()
+  const output = _.pipeline(stream => {
+    const temp = _()
+    stream
+    .pipe(vrt)
+    .on('finish', () => {
+      if (format === 'zip') return shapefile.createStream(options)
+      const ogr = spawn('ogr2ogr', cmd)
+      // TODO can I just pipe out vs writing to temp?
+      ogr.stdout.on('data', data => temp.write(data))
+      ogr.stderr.on('data', data => output.emit('log', {level: 'debug', message: data.toString()}))
+      ogr.on('close', c => {
+        output.emit('log', {level: 'info', message: `Executing: OGR2OGR ${cmd.join(' ')}`})
+        if (c > 0) output.emit('error', new Error('OGR Failed'))
+        else temp.end()
+      })
     })
-
-    return readStream
-  },
-  /**
-   * Gets a set of OGR Parameters for an export
-   *  
-   * @param {object} options - potentially contains a fields object
-   */
-  createCmd: function (options) {
-    const ogrFormat = ogrFormats[options.format]
-    // shapefiles cannot be streamed out of ogr2ogr
-    const output = options.format === 'zip' ? `${options.path || '.'}/${options.name}` : '/vsistdout/'
-    const input = options.input || 'layer.vrt'
-
-    let cmd = ['--config', 'SHAPE_ENCODING', 'UTF-8', '-f', ogrFormat, output, input]
-
-    options.geometryType = options.geometryType && options.geometryType.toUpperCase() || 'NONE'
-    if (options.format === 'csv') cmd = csvParams(cmd, options)
-    if (options.format === 'zip') cmd = shapefileParams(cmd, options)
-    return finishOgrParams(cmd)
-  }
+    return temp
+  })
+  return output
 }
 
-/**
- * Add parameters specific to a csv export
- *
- * @param {array} cmd - an array of OGR command parts to modify
- * @param {object} options - may contain fields
- * @return {array}
- * @private
- */
-function csvParams (cmd, options) {
-  cmd.push('-lco')
-  cmd.push('WRITE_BOM=YES')
-  const hasPointGeom = options.geometryType === 'POINT'
-  const fields = options.fields.join('|').toLowerCase().split('|')
-  const hasXY = fields.indexOf('x') > -1 && fields.indexOf('y') > -1
-  if (hasPointGeom && !hasXY) cmd = cmd.concat(['-lco', 'GEOMETRY=AS_XY'])
-  return cmd
-}
-
-/**
- * Add parameters specific to a csv export
- *
- * @param {string} cmd - an array of OGR command parts to modify
- * @param {object} options - may contain a wkid or wkt
- * @param {function} callback - calls back back with a modified command array or an error
- * @private
- */
-function shapefileParams (cmd, options) {
-  // make sure geometries are still written even if the first is null
-  cmd.push(`-nlt ${options.geometryType.toUpperCase()}`)
-  cmd.push('-fieldmap')
-  cmd.push('identity')
-  if (options.wkt) cmd.push(`-t_srs '${options.wkt}'`)
-  return cmd
-}
-
-/**
- * Adds final parameters to the OGR function call
- * @param {array} cmd - an array of OGR command parts to modify
- * @return {string} the final OGR command
- */
-function finishOgrParams (cmd) {
-  cmd.push('-append')
-  cmd.push('-skipfailures')
-  cmd.push('-lco')
-  cmd.push('ENCODING=UTF-8')
-  return cmd
-}
+module.exports = {createStream}