+++ test/csv.js
@@ -1,6 +1,7 @@
 const test = require('tape')
 const Ogr = require('../src/lib/ogr.js')
 const Helper = require('./helper')
+const _ = require('highland')
 
 test('Set up', t => {
   Helper.before()
@@ -11,30 +12,32 @@ test('Create a csv readstream', t => {
   t.plan(1)
   const options = defaultOptions()
   const rows = []
-  Ogr.createReadStream(Helper.testVrt, options)
-    .on('error', e => {
-      console.log(e); t.end(e)
-    })
-    .split()
-    .compact()
-    .each(row => {
-      rows.push(row)
-    })
-    .done(() => {
-      t.equal(rows.length, 101, 'All rows written to the stream')
-    })
+  Helper.vrt()
+  .pipe(Ogr.createStream('csv', options))
+  .on('error', e => {
+    console.log(e); t.end(e)
+  })
+  .split()
+  .compact()
+  .each(row => {
+    rows.push(row)
+  })
+  .done(() => {
+    t.equal(rows.length, 101, 'All rows written to the stream')
+  })
 })
 
-test('Gracefully handle a vrt that does not exist', t => {
+test('Gracefully handle a malformed VRT', t => {
   t.plan(1)
   const options = defaultOptions()
   try {
-    Ogr.createReadStream('foobar!', options)
-      .on('error', err => {
-        t.ok(err, 'Error was caught in the correct place')
-        t.end()
-      })
-      .done(() => t.end())
+    _(['Foobar!'])
+    .pipe(Ogr.createStream('csv', options))
+    .on('error', err => {
+      t.ok(err, 'Error was caught in the correct place')
+      t.end()
+    })
+    .done(() => t.end())
   } catch (e) {
     t.fail('Error was uncaught')
   }
@@ -47,9 +50,9 @@ test('Teardown', t => {
 
 function defaultOptions () {
   return {
+    path: `${__dirname}/output`,
     name: 'dummy',
-    format: 'csv',
-    geometryType: 'Point',
+    geometry: 'Point',
     fields: ['foo', 'bar', 'baz']
   }
 }