+++ test/kml.js
@@ -1,6 +1,7 @@
 const test = require('tape')
 const Ogr = require('../src/lib/ogr.js')
 const Helper = require('./helper')
+const _ = require('highland')
 
 test('Set up', t => {
   Helper.before()
@@ -11,7 +12,8 @@ test('Create a kml readstream', t => {
   t.plan(1)
   const options = defaultOptions()
   const rows = []
-  Ogr.createReadStream(Helper.testVrt, options)
+  Helper.vrt()
+  .pipe(Ogr.createStream('kml', options))
   .on('error', e => {console.log(e); t.end(e)})
   .splitBy('<Placemark>')
   .compact()
@@ -23,15 +25,16 @@ test('Create a kml readstream', t => {
    })
 })
 
-test('Gracefully handle a vrt that does not exist', t => {
+test('Gracefully handle a malformed VRT', t => {
   t.plan(1)
   const options = defaultOptions()
   try {
-    Ogr.createReadStream('foobar!', options)
+    _(['Foobar!'])
+    .pipe(Ogr.createStream('kml', options))
     .on('error', err => {
       t.ok(err, 'Error was caught in the correct place')
       t.end()
-     })
+    })
     .done(() => t.end())
   } catch (e) {
     t.fail('Error was uncaught')
@@ -45,8 +48,8 @@ test('Teardown', t => {
 
 function defaultOptions () {
   return {
+    path: `${__dirname}/output`,
     name: 'dummy',
-    format: 'kml',
-    geometryType: 'Point'
+    geometry: 'Point'
   }
 }