+++ test/ogr-cmd.js
@@ -1,16 +1,15 @@
 const test = require('tape')
-const Ogr = require('../src/lib/ogr')
+const Cmd = require('../src/lib/ogr-cmd')
 
 test('Create a command for a csv when there are x y properties', t => {
   t.plan(1)
   const options = {
     name: 'dummy',
-    format: 'csv',
     fields: ['foo', 'bar', 'X', 'y'],
-    geometryType: 'Point'
+    geometry: 'Point'
   }
 
-  const cmd = Ogr.createCmd(options)
+  const cmd = Cmd.create('csv', options)
   t.equal(cmd.join(' '), '--config SHAPE_ENCODING UTF-8 -f CSV /vsistdout/ layer.vrt -lco WRITE_BOM=YES -append -skipfailures -lco ENCODING=UTF-8', 'Correct command')
 })
 
@@ -20,10 +19,10 @@ test('Create a command for a csv when there are not x or y features', t => {
     name: 'dummy',
     format: 'csv',
     fields: ['foo', 'bar'],
-    geometryType: 'Point'
+    geometry: 'Point'
   }
 
-  const cmd = Ogr.createCmd(options)
+  const cmd = Cmd.create('csv', options)
   t.equal(cmd.join(' '), '--config SHAPE_ENCODING UTF-8 -f CSV /vsistdout/ layer.vrt -lco WRITE_BOM=YES -lco GEOMETRY=AS_XY -append -skipfailures -lco ENCODING=UTF-8', 'Correct command')
 })
 
@@ -33,10 +32,10 @@ test('Create a command for a csv when there is no geometry', t => {
     name: 'dummy',
     format: 'csv',
     fields: ['foo', 'bar'],
-    geometryType: null
+    geometry: null
   }
 
-  const cmd = Ogr.createCmd(options)
+  const cmd = Cmd.create('csv', options)
   t.equal(cmd.join(' '), '--config SHAPE_ENCODING UTF-8 -f CSV /vsistdout/ layer.vrt -lco WRITE_BOM=YES -append -skipfailures -lco ENCODING=UTF-8', 'Correct command')
 })
 
@@ -44,23 +43,21 @@ test('Create a command for a shapefile', t => {
   t.plan(1)
   const options = {
     name: 'dummy',
-    format: 'zip',
-    geometryType: 'Point'
+    geometry: 'Point'
   }
 
-  const cmd = Ogr.createCmd(options)
+  const cmd = Cmd.create('zip', options)
   t.equal(cmd.join(' '), '--config SHAPE_ENCODING UTF-8 -f "ESRI Shapefile" ./dummy layer.vrt -nlt POINT -fieldmap identity -append -skipfailures -lco ENCODING=UTF-8', 'Correct command')
 })
 
-test('Create the command for a shapefile when a wkt is passed in', t => {
+test('Create the command for a shapefile when a srs is passed in', t => {
   t.plan(1)
   const options = {
     name: 'dummy',
-    wkt: 'PROJECTION',
-    format: 'zip',
-    geometryType: 'Point'
+    srs: 'PROJECTION',
+    geometry: 'Point'
   }
 
-  const cmd = Ogr.createCmd(options)
+  const cmd = Cmd.create('zip', options)
   t.equal(cmd.join(' '), '--config SHAPE_ENCODING UTF-8 -f "ESRI Shapefile" ./dummy layer.vrt -nlt POINT -fieldmap identity -t_srs \'PROJECTION\' -append -skipfailures -lco ENCODING=UTF-8', 'Correct command')
 })